Progress: Begin implementation of the 'remove' series of programs.
Much of the work is drafted out and likely to change.

The core of the 'remove' program is implemented.
The 'date' related functionality is actively being worked on.

The actual removal functionality will be implemented after all of the simulation is completed.
Once the remove is implemented and working, then supplementary 'remove' programs, such as 'rm', 'unlink', and 'rmdir' will be implemented.