Progress: Continue working on remove program.
Significant development towards completing date and time related checks.
This is not yet complete.

Fix existence test mistake, do not treat the file not found error as an error during simulation.

The settings loading has several loops.
Redesign the main function to start the thread.
The settings loading now checks in some of its loops for an interrupt signal.