Progress: Continue work.
Add incomplete copyright printing.

Remove or disable the /proc relating "isolate" functions.
I will think later on if and how I want to support that.
It may be usefully to have a "--proc=/proc" style long parameter if this is decided to be kept.

Add some needed changes that I missed in the previous commit regarding updating to FLL 0.7.x/0.8.x.

Generalize some of the functions a bit more than previously done.

Add more comments about plans to make the future work easier.
It took me a bit of time to remember where I left off and these comments should now help me in the future.