Progress: Miscellaneous improvements to the Remove program.
Update the build settings file to be more in line with the FLL program build settings files.

Update the structure where I noticed it has become out of synchronization with the FLL programs structure.
Fix minor mistakes such as "utf8" is present in the file when it should instead be "remove".