Feature: Initial addition of the "TacocaT" program.
The "TacocaT" program is intended to be similar to netcat.
To concatenate multiple network connections in this context simply means to transfer data from one point to another.

The intent and purpose of this program is really to test my network connection software.

This program is only stubbed out at this point in time.

I need to figure out how I want the data files to be setup.
The network code is not implemented at all, yet.

I also decided to tease my plans on improving the existing Kevux DNS alternative network resolution methodology.