Progress: Further work in getting TacocaT working.
I decided to keep things simple and remove the background process and PID file related functionality.

I renamed "from" and "to" into "receive" and "send".
This relating context makes more sense and can be less confusing.

The network and the local files should be just about ready to use.
The actual implementation of the network processing code is next.

I intend to use either threading or forking to handle each send or receive connection set.