Progress: Further work in TacocaT, getting the "receive" working and writing to a file.
This gets the receive mostly working.
There may be a few fixme or todo comments to address.
I decided to not bother with some of the more complex handling of problems to get this program out the door.

To test this, I used the FLL-0.6 branch version of the file specifications/fss-000f.txt.
I then created the Packet header using the bash script like this:
  printf '\x00\x00\x00\x15\x78' > test.binary
  cat 0.6/specifications/fss-000f.txt >> test.binary

The resulting binary file size is 0x1578 bytes (5496 bytes) and this can be confirmed via the wc command:
  wc test.binary

I can then start the TacocaT program in a new terminal like this:
  tacocat -r 127.0.0.1:12345 test.out +V

In a different terminal, I use netcat to test (because the "send" support is not yet written).
  cat test.binary | netcat 127.0.0.1 12345

Use control-c to terminate TacocaT.

I can then confirm that the file has been created successfully via a md5sum check.
  md5sum test.out 0.6/specifications/fss-000f.txt

The checksums should match, and in my testing they do match.