Progress: Further work in TacocaT.
The socket id and id_data descriptors need to be initialized to -1 and not 0.
Make sure they are setup after allocation and after processing (or after disconnect).

Provide more verbose error reporting that reports the particular block step the send fails on.
The receive is not yet updated to do the same but in the future it should be updated to do so.