Progress: Further work in TacocaT.
I previously just jammed on a random seed based on time as a quick short term solution.

I created an f_random FLL project to provide some basic random entropy processing code.

This change set focuses on utilizing these to get set the random entropy.

The setup of the seed is set to non-blocking with a single delayed retry if blocked.
On failure, this still falls back to the time() based setup.
The fallback also must be cast t an unsigned int rather than a long.