Progress: Further work in TacocaT.
This strongly moves in the direction of setting up sending and receiving the different packets.
This includes handling of multiple packet parts.

None of this is ready yet, but progress is being made.
There is currently a strong focus on getting the FSS Payload processing code wrapped up, working, and tested.