Progress: Further work in TacocaT, updating "salt" and "time" packet headers.
The "salt" header is now in the FLL project.
Remove the "salt" string definition from this project.

The packet send now utilizes the "time" header via the standard UTC time.
This is for informative purposes rather than for any accurancy.
Which means that the lack of microseconds or nanoseconds is acceptable.